% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_windows.R
\name{create_corridors}
\alias{create_corridors}
\title{Creates}
\usage{
create_corridors(roads, fclass = "fclass", name = "name")
}
\arguments{
\item{roads}{roads simple feature object}

\item{fclass}{title of the attribute with the fclass values}

\item{name}{title of the attribute with the street names}
}
\value{
a simple features object of corridors
}
\description{
Results are road segments aggregated by name, functional class, and proximity to each other. Can be considered corridors.
}
