% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_windows.R
\name{create_sliding_windows}
\alias{create_sliding_windows}
\title{Builds Sliding Windows Analysis}
\usage{
create_sliding_windows(
  roads,
  name = "name",
  fclass = "fclass",
  win = 0.5,
  step = 0.1
)
}
\arguments{
\item{roads}{Path to the input file}

\item{name}{title of the attribute with the street names}

\item{fclass}{title of the attribute with the fclass values}

\item{win}{max window length in miles}

\item{step}{distance between each step in miles}
}
\value{
a simple features object of windows from sliding windows analysis
}
\description{
Results are a polyline object that can help to identify locations 
with a high frequency of crashes, highlighting potential high-injury areas.
}
